Bugfix HeightProfileForSlice and fixed intersect
Bugfix in HeightProfileForSlice. Estimation of outside part at start was going wrong. I think (!) I fixed it, but really hard to test. Also, not really relevant.
Also, for sector centers outside the image, the intersect with the image is now fixed. GetIntersectsImageBorder does not solely rely now on polangle (which is equal for centers outside image), but now includes the length for these slices.